requirements.txt
Django==4.2.7
djangorestframework==3.14.0
cryptography==41.0.5
python-dateutil==2.8.2
PyJWT==2.8.0
rich==13.6.0